// src/types/generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  Dimension: { input: any; output: any; }
  HexColor: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Quality: { input: any; output: any; }
};
// ...
export type AssetCollection = {
  __typename?: 'AssetCollection';
  items: Array<Maybe<Asset>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type ContentfulMetadata = {
  __typename?: 'ContentfulMetadata';
  concepts: Array<Maybe<TaxonomyConcept>>;
  tags: Array<Maybe<ContentfulTag>>;
};
// ...
export type ContentfulTag = {
  __typename?: 'ContentfulTag';
  id?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
};

/** [See type definition](https://app.contentful.com/spaces/gyfunrv4a4ak/content_types/customer) */
export type Customer = Entry & _Node & {
  __typename?: 'Customer';
  _id: Scalars['ID']['output'];
  contentfulMetadata: ContentfulMetadata;
  firstName?: Maybe<Scalars['String']['output']>;
  lastName?: Maybe<Scalars['String']['output']>;
  linkedFrom?: Maybe<CustomerLinkingCollections>;
  sys: Sys;
};
// ...
export type CustomerLinkingCollections = {
  __typename?: 'CustomerLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
  reviewCollection?: Maybe<ReviewCollection>;
};
// ...
export type Entry = {
  contentfulMetadata: ContentfulMetadata;
  sys: Sys;
};

export type EntryCollection = {
  __typename?: 'EntryCollection';
  items: Array<Maybe<Entry>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type Hotel = Entry & _Node & {
  __typename?: 'Hotel';
  _id: Scalars['ID']['output'];
  city?: Maybe<Scalars['String']['output']>;
  contentfulMetadata: ContentfulMetadata;
  country?: Maybe<Scalars['String']['output']>;
  description?: Maybe<HotelDescription>;
  endDate?: Maybe<Scalars['DateTime']['output']>;
  imagesCollection?: Maybe<AssetCollection>;
  linkedFrom?: Maybe<HotelLinkingCollections>;
  name?: Maybe<Scalars['String']['output']>;
  price?: Maybe<Scalars['JSON']['output']>;
  rating?: Maybe<Scalars['Int']['output']>;
  startDate?: Maybe<Scalars['DateTime']['output']>;
  sys: Sys;
};
// ...
export type HotelDescription = {
  __typename?: 'HotelDescription';
  json: Scalars['JSON']['output'];
  links: HotelDescriptionLinks;
};
// ...
export type HotelLinkingCollections = {
  __typename?: 'HotelLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
  reviewCollection?: Maybe<ReviewCollection>;
};
// ...
export type Review = Entry & _Node & {
  __typename?: 'Review';
  _id: Scalars['ID']['output'];
  comment?: Maybe<ReviewComment>;
  contentfulMetadata: ContentfulMetadata;
  customer?: Maybe<Customer>;
  feedback?: Maybe<Scalars['String']['output']>;
  hotel?: Maybe<Hotel>;
  linkedFrom?: Maybe<ReviewLinkingCollections>;
  rating?: Maybe<Scalars['Int']['output']>;
  sys: Sys;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type ReviewComment = {
  __typename?: 'ReviewComment';
  json: Scalars['JSON']['output'];
  links: ReviewCommentLinks;
};

export type ReviewCommentAssets = {
  __typename?: 'ReviewCommentAssets';
  block: Array<Maybe<Asset>>;
  hyperlink: Array<Maybe<Asset>>;
};

export type ReviewCommentEntries = {
  __typename?: 'ReviewCommentEntries';
  block: Array<Maybe<Entry>>;
  hyperlink: Array<Maybe<Entry>>;
  inline: Array<Maybe<Entry>>;
};

export type ReviewCommentLinks = {
  __typename?: 'ReviewCommentLinks';
  assets: ReviewCommentAssets;
  entries: ReviewCommentEntries;
  resources: ReviewCommentResources;
};
// ...
export type ReviewLinkingCollections = {
  __typename?: 'ReviewLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
};
// ...
export type Sys = {
  __typename?: 'Sys';
  environmentId: Scalars['String']['output'];
  firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  id: Scalars['String']['output'];
  /** The locale that was requested. */
  locale?: Maybe<Scalars['String']['output']>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  publishedVersion?: Maybe<Scalars['Int']['output']>;
  spaceId: Scalars['String']['output'];
};
// ...
export type TaxonomyConcept = {
  __typename?: 'TaxonomyConcept';
  id?: Maybe<Scalars['String']['output']>;
};

export type _Node = {
  _id: Scalars['ID']['output'];
};